Skip merge accessors before combine nodes/meshes
Running merge accessors before nodes/meshes when multiple accessors can
be independent but have the same data causes one of the accessors to be
deleted and both meshes to use the same accessor. When this happens,
combineNodes can no longer merge the meshes.

Skipping merge accessors allows the nodes to be merged first and the
accessors can be merged during the second call to
mergeDuplicateProperties.