Add option to output uncompressed fallback mesh
If the uncompressed fallback option is set, gltf-pipeline will output compressed meshes as well as uncompressed meshes. The uncompressed mesh and the compressed mesh will be written to their own files, so rendering apps do not have to download both types of meshes.
If the uncompressed fallback option is set, extension required will not be written.
If only compressed meshes are output, indices and attribute counts in accessors will not be written.